requirements.txt
tushare
retrying==1.3.3
pytz
mock
sentry_sdk
python_dateutil==2.8.0
pandas>=0.18.0
requests>=2.0.0
pandas_datareader>=0.4.0
Numpy>=1.11.1
lxml>=4.5.1
tornado>=6.0
awswrangler
pynamodb==4.3.3